backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
